Restructure package.
